refactor: extract database into a standalone file
